+++ ios/Classes/NotificationDetails.h
@@ -11,6 +11,7 @@
 @property(nonatomic) bool presentBadge;
 @property(nonatomic, strong) NSString *sound;
 @property(nonatomic, strong) NSNumber *secondsSinceEpoch;
+@property(nonatomic, strong) NSNumber *scheduleDate;
 @property(nonatomic, strong) NSNumber *repeatInterval;
 @property(nonatomic, strong) NotificationTime *repeatTime;
 @property(nonatomic, strong) NSNumber *day;